add rbsi in opron, and delete old_operon
