--- src/apps/operon_old/details/info.module.css
@@ -1,17 +1,0 @@
-.tabHeader {
-  width: 100%;
-  padding-left: 10%;
-  display: flow-root;
-}
-
-.tabContent {
-  float: left;
-  padding-right: 1%;
-  font-size: 12px;
-}
-
-.tab_type{
-  color: #666666;
-  font-weight: normal;
-  font-size: 10px;
-}
