--- src/apps/operon_old/details/summary/index.jsx
@@ -1,9 +1,0 @@
-function Summary() {
-    return ( 
-        <div>
-            Hola
-        </div>
-     );
-}
-
-export default Summary;
