--- src/apps/operon_old/operon.css
@@ -1,11 +1,0 @@
-.cover_statistics{
-    display: flex;
-}
-
-.stt_box{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: var(--color-blue5);
-    margin: 0 2px;
-}
