+++ src/components/datamartSchema/operon/transcriptionUnit/index.jsx
@@ -2,11 +2,10 @@ import DrawingTracesTool from "../../../DrawingTracesTool"
 import Genes from "./genes"
 import Promoter from "./promotor";
 import Divider from "@mui/material/Divider";
+import Terminators from "./terminators";
 
 import { Accordion, DataVerifier } from "../../../ui-components"
-import { ParagraphCitations } from "../../citations";
-import { LinealSequence } from "../../../sequence";
-import { useMemo } from "react";
+import RegulatorBindingSites from "./regulatorBindingSites";
 
 
 export default function TranscriptionUnit({
@@ -27,6 +26,10 @@ export default function TranscriptionUnit({
     synonyms = [],
     terminators = []
 }) {
+    let title = "Detailed TU " + name
+    if (DataVerifier.isValidObject(promoter)) {
+        title = title + " and promoter "+promoter.name
+    }
 
 
     return (
@@ -44,10 +47,10 @@ export default function TranscriptionUnit({
                 />
             </div>
             <div>
-                <Accordion title={"Detailed TU " + name + " with promoter " + promoter.name + " information"} backgroundColor="#ffffff" >
+                <Accordion title={title } backgroundColor="#ffffff" >
                     <div>
                         {DataVerifier.isValidObject(firstGene) && (
-                            <Genes firstGene={firstGene} genes={genes} />
+                            <Genes allCitations={allCitations} firstGene={firstGene} genes={genes} />
                         )}
                         <Divider />
                         {DataVerifier.isValidObject(promoter) && (
@@ -57,6 +60,10 @@ export default function TranscriptionUnit({
                         {DataVerifier.isValidArray(terminators) && (
                             <Terminators terminators={terminators} tuID={_id} allCitations={allCitations} />
                         )}
+                        <Divider />
+                        {DataVerifier.isValidArray(regulatorBindingSites) && (
+                            <RegulatorBindingSites regulatorBindingSites={regulatorBindingSites} allCitations={allCitations} />
+                        )}
                     </div>
                 </Accordion>
             </div>
@@ -64,81 +71,3 @@ export default function TranscriptionUnit({
     )
 }
 
-function Terminators({ terminators, tuID, allCitations }) {
-    return (
-        <div>
-            <h2>Terminators</h2>
-            <div style={{ marginLeft: "5px", display: "flex", flexDirection: "column" }}>
-                {terminators.map((terminator, index) => (<Terminator key={`T${index}_tu_${tuID}_terminator${terminator._id}`} allCitations={allCitations} terminator={terminator} />))}
-            </div>
-        </div>
-    )
-}
-
-function Terminator({ terminator, allCitations }) {
-    let _confidenceLevel
-    if (DataVerifier.isValidString(terminator.confidenceLevel)) {
-        switch (terminator.confidenceLevel) {
-            case "S":
-                _confidenceLevel = <span style={{ fontWeight: "bold", color: "#0C6A87" }} >Strong</span>
-                break;
-            case "C":
-                _confidenceLevel = <span style={{ fontWeight: "bold", color: "#000000" }} >Confirmed</span>
-                break;
-            case "w":
-                _confidenceLevel = <span style={{ color: "#0C6A87" }} >Weak</span>
-                break;
-            default:
-                _confidenceLevel = <span>.</span>
-                break;
-        }
-    }
-
-    return (
-        <div>
-            {DataVerifier.isValidString(terminator.class) && (
-                <p><b>Class:</b>{" " + terminator.class}</p>
-            )}
-            {DataVerifier.isValidString(terminator.confidenceLevel) && (
-                        <p><b>Confidence Level:</b>{" "}{_confidenceLevel}</p>
-                    )}
-            {DataVerifier.isValidObject(terminator.transcriptionTerminationSite) && (
-                <p><b>Transcription Site:</b>{" " + terminator.transcriptionTerminationSite.leftEndPosition + " - " + terminator.transcriptionTerminationSite.rightEndPosition}</p>
-            )}
-            {DataVerifier.isValidString(terminator.sequence) && (
-                <Accordion title={"sequence"} >
-                    <SequenceTerminator _id={terminator._id} transcriptionTerminationSite={terminator.transcriptionTerminationSite} sequence={terminator.sequence} name={`Terminator_${terminator._id}`} />
-                </Accordion>
-            )}
-            {DataVerifier.isValidArray(terminator.citations) && (
-                <p><b>Citations:</b><br /><ParagraphCitations citations={terminator.citations} allCitations={allCitations} /></p>
-            )}
-            <br />
-            <br />
-        </div>
-    )
-}
-
-function SequenceTerminator({ _id, sequence, name, transcriptionTerminationSite }) {
-
-    const features = useMemo(() => {
-        let _features = []
-        const terminatorLength = Math.abs(transcriptionTerminationSite.rightEndPosition - transcriptionTerminationSite.leftEndPosition)
-        const terminatorInit = sequence.split("").findIndex(bp => bp === bp.toUpperCase())
-        _features.push({
-            id: _id + "_terminator__feature",
-            sequencePosition: terminatorInit,
-            length: terminatorLength,
-            type: "terminator",
-        })
-        return _features
-    }, [_id, sequence, transcriptionTerminationSite])
-
-
-
-
-    return <LinealSequence name={name} sequenceId={_id} height={100} sequence={sequence} color={true} features={features} />
-
-}
-
-//gtggattatgTTCAGCGCGAGCTGGCAGACGGTAGCCGTACCGTTGTCGAAACCGAACACTGGTTAGCCGTCGTGCCTTACTGGGCTGCCTGGCCGTTCGAAACGCTACTGCTGCCCAAAGCCCacgttttacg
