+++ src/components/filterTable/table.module.css
@@ -13,7 +13,7 @@
 
 .resizer {
     display: inline-block;
-    background: #72a7c7;
+    background: #FFFFFF;
     width: 2px;
     height: 100%;
     position: absolute;
@@ -30,7 +30,7 @@
 
 .resizerRight {
     display: inline-block;
-    background: #72a7c7;
+    background: #FFFFFF;
     color:#72a7c7;
     width: 2px;
     height: 100%;
